backend/requirements.txt
Django==5.0.7
django-ninja==1.1.0
daphne==4.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
redis==5.0.6
celery==5.4.0
pydantic~=2.7.4
boto3~=1.34.131
botocore~=1.34.131
pillow==10.3.0
pillow_heif==0.16.0
piexif==1.1.3
requests==2.32.3
channels==4.1.0
channels-redis==4.2.0
asgiref~=3.8.1